requirements.txt
# StudyBuddyAI - AI 作业监督系统
# Python >= 3.11

# Web 框架
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0

# Google Gemini Live API
google-genai>=1.0.0

# 火山引擎豆包 API (备选)
volcengine-python-sdk>=1.0.0

# 数据处理
pydantic>=2.10.0
pyyaml>=6.0.2

# 异步支持
aiofiles>=24.1.0
httpx>=0.28.0

# 音频处理
pyaudio>=0.2.14

# 图像处理
pillow>=11.0.0
opencv-python-headless>=4.10.0

# 日志
structlog>=24.4.0

# 环境变量
python-dotenv>=1.0.1

# 开发工具
ruff>=0.8.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
